// src/app/shared-module/enum/operator.enum.ts
/**
 * 操作符枚举类
 */
export enum OperatorEnum {
  /**
   * like模糊查询
   */
  like = 'like',
  /**
   * 等于
   */
  eq = 'eq',
  /**
   * in在里面
   */
  in = 'in',
  /**
   * gte
   */
  gte = 'gte',
  /**
   * lte
   */
  lte = 'lte',
  /**
   * gt大于
   */
  gt = 'gt',
  /**
   * lt小于
   */
  lt = 'lt'
}